Compute investment growth with annuities

This function takes an initial investment amount, an amount for annuities added to the investment every year, an annual return rate, and years of investment, and returns a numeric vector representing the new amount every year.

Usage

annuity(return_rate, initial_investment, annuity, years)

Arguments

return_rate

A numeric representing the annual return rate.

initial_investment

A numeric representing the initial investment amount.

annuity

A numeric representing the amount added to the investment every year.

years

A numeric representing the number of years of the investment period.

Value

A numeric vector representing the size of the investment each year.

Examples

# With an initial investment of $10,000 and yearly contributions of $5,000 ...
# ... what will the size of the investment be after 25 years of 8% annual returns?
annuity(.08,10000,5000,25)
#>  [1]  10800.00  17064.00  23829.12  31135.45  39026.29  47548.39  56752.26
#>  [8]  66692.44  77427.84  89022.06 101543.83 115067.33 129672.72 145446.54
#> [15] 162482.26 180880.84 200751.31 222211.41 245388.33 270419.39 297452.94
#> [22] 326649.18 358181.11 392235.60 429014.45