Compute your net income from your gross income, after factoring in federal and state taxes

This function takes a numeric value representing your gross income and returns a tibble with information like net income, the amount you pay in federal taxes and state taxes (possible states are Oklahoma and California), and other things. Married filing jointly is assumed.

Usage

taxes(gross_income, state_name = "OK")

Arguments

gross_income

A numeric positive value representing your gross annual income.

state_name

A character. 'OK' for Oklahoma, and 'CA' for California. Default is 'OK'.

Value

A tibble with the following things: Your federal top marginal rate, state top marginal rate, federal taxes, state taxes, amount paid to social security and medicare, proportion removed from gross income, net income, and what your net income would be if there were no state taxes.

Examples

# What is your net income in Oklahoma if your gross income is $80,000?
taxes(80000)
#> # A tibble: 1 × 10
#>   Gross_Income Tax_Filing Fed_Top_Marginal   State_Top_Marginal   Fed_Taxes
#>          <dbl> <chr>      <chr>              <chr>                    <dbl>
#> 1        80000 Married    0.12, 19901, 0.751 0.0525, 15001, 0.812     9202.
#> # ℹ 5 more variables: State_Taxes <dbl>, SocialSec_Medicare <chr>,
#> #   Prop_Taken <dbl>, Net_Income <dbl>, If_No_State_Tax <dbl>